+++ src/main.ts
@@ -4,11 +4,7 @@ import {
   withInterceptorsFromDi,
 } from '@angular/common/http';
 import { APP_INITIALIZER, importProvidersFrom, isDevMode } from '@angular/core';
-import {
-  bootstrapApplication,
-  HammerModule,
-  HAMMER_GESTURE_CONFIG,
-} from '@angular/platform-browser';
+import { bootstrapApplication } from '@angular/platform-browser';
 import { provideAnimations } from '@angular/platform-browser/animations';
 import { RouterModule } from '@angular/router';
 import { DateAdapter, NativeDateAdapter } from '@angular/material/core';
@@ -16,7 +12,6 @@ import { MatIconRegistry } from '@angular/material/icon';
 import { TranslocoService, provideTransloco } from '@ngneat/transloco';
 import { provideTranslocoPersistLang } from '@ngneat/transloco-persist-lang';
 import { ToastrModule } from 'ngx-toastr';
-import 'hammerjs';
 
 import { AuthInterceptor } from '@interceptors/auth.interceptor';
 import { ErrorHandlerInterceptor } from '@interceptors/error-handler.interceptor';
@@ -24,7 +19,6 @@ import { TranslocoHttpLoader, preloadLanguage } from './transloco-loader';
 import { AppComponent } from './app/app.component';
 import { routes } from './app/routes';
 import { BACKEND_URL, FRONTEND_URL } from './tokens';
-import { LTHammerConfig } from './config.hammerjs';
 import {
   backendUrlDevelopment,
   backendUrlProduction,
@@ -53,8 +47,6 @@ bootstrapApplication(AppComponent, {
     { provide: DateAdapter, useClass: NativeDateAdapter },
     provideAnimations(),
     provideHttpClient(withInterceptorsFromDi()),
-    importProvidersFrom(HammerModule),
-    { provide: HAMMER_GESTURE_CONFIG, useClass: LTHammerConfig },
     importProvidersFrom(RouterModule.forRoot(routes)),
     importProvidersFrom(ToastrModule.forRoot()),
     provideTranslocoPersistLang({